[mesh] extract mvsUtils::createRefMeshFromDenseSfMData from main_texturing
